+++ examples/parsing_graphs/src/auxiliary_functions.jl
@@ -40,33 +40,5 @@ function get_deps_promises(vertex_id, map, G)
     return incoming_data
 end
 
-function parse_graphml(path)
-    file_path = joinpath(path...)
-    G = GraphMLReader.loadgraphml(file_path, "G")
-end
 
-function show_graph(G)
-    for (_, v) in enumerate(Graphs.vertices(G))
-        println("Node: ")
-        print("Node type: ")
-        println(get_prop(G, v, :type))
-        if has_prop(G, v, :class)
-            print("Node class: ")
-            println(get_prop(G, v, :class))
-        end
-        if has_prop(G, v, :runtime_average_s)
-            print("Average runtime [s]: ")
-            println(get_prop(G, v, :runtime_average_s))
-        end
-        if has_prop(G, v, :size_average_B)
-            print("Average size [B]: ")
-            println(get_prop(G, v, :size_average_B))
-        end
-        print("Original name: ")
-        println(get_prop(G, v, :original_id))
-        print("Node name: ")
-        println(get_prop(G, v, :node_id))
-        println()
-    end
-end
